+++ src/bin/zarrs_benchmark_read_sync.rs
@@ -1,8 +1,19 @@
-use std::{sync::Arc, sync::Mutex, time::SystemTime};
+use std::{
+    sync::{Arc, Mutex},
+    time::SystemTime,
+};
 
 use clap::Parser;
-use rayon::iter::{ParallelBridge, ParallelIterator};
-use zarrs::{array_subset::ArraySubset, storage::ReadableStorageTraits};
+use rayon::iter::{IntoParallelIterator, ParallelIterator};
+use zarrs::{
+    array::{
+        codec::{ArrayCodecTraits, CodecOptionsBuilder},
+        concurrency::RecommendedConcurrency,
+    },
+    array_subset::ArraySubset,
+    config::global_config,
+    storage::ReadableStorageTraits,
+};
 
 #[derive(Parser, Debug)]
 #[command(
@@ -16,8 +27,8 @@ struct Args {
     path: String,
 
     /// Number of concurrent chunks.
-    #[arg(long, default_value_t = 4)]
-    concurrent_chunks: usize,
+    #[arg(long)]
+    concurrent_chunks: Option<usize>,
 
     /// Read the entire array in one operation.
     ///
@@ -38,7 +49,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         args.path.clone(),
     )?);
     let array = zarrs::array::Array::new(storage.clone(), "/")?;
-    println!("{:#?}", array.metadata());
+    // println!("{:#?}", array.metadata());
 
     zarrs::config::global_config_mut().set_validate_checksums(!args.ignore_checksums);
 
@@ -48,21 +59,50 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let bytes_decoded = Mutex::new(0);
     if args.read_all {
         let subset = ArraySubset::new_with_shape(array.shape().to_vec());
-        *bytes_decoded.lock().unwrap() += array.par_retrieve_array_subset(&subset)?.len();
+        *bytes_decoded.lock().unwrap() += array.retrieve_array_subset(&subset)?.len();
     } else {
-        (0..chunks.shape().iter().product())
-            .collect::<Vec<_>>()
-            .as_slice()
-            .chunks((chunks.num_elements_usize() + args.concurrent_chunks - 1) / args.concurrent_chunks)
-            .par_bridge()
-            .for_each(|chunk_index_chunk| {
-                for chunk_index in chunk_index_chunk {
-                    let chunk_indices = zarrs::array::unravel_index(*chunk_index, chunks.shape());
-                    // println!("Chunk/shard: {:?}", chunk_indices);
-                    let bytes = array.retrieve_chunk(&chunk_indices).unwrap();
-                    *bytes_decoded.lock().unwrap() += bytes.len();
-                }
-            });
+        let chunk_representation =
+            array.chunk_array_representation(&vec![0; array.chunk_grid().dimensionality()])?;
+        let concurrent_target = std::thread::available_parallelism().unwrap().get();
+        let (chunks_concurrent_limit, codec_concurrent_target) =
+            zarrs::array::concurrency::calc_concurrency_outer_inner(
+                concurrent_target,
+                &if let Some(concurrent_chunks) = args.concurrent_chunks {
+                    let concurrent_chunks =
+                        std::cmp::min(chunks.num_elements_usize(), concurrent_chunks);
+                    RecommendedConcurrency::new(concurrent_chunks..concurrent_chunks)
+                } else {
+                    let concurrent_chunks = std::cmp::min(
+                        chunks.num_elements_usize(),
+                        global_config().chunk_concurrent_minimum(),
+                    );
+                    RecommendedConcurrency::new_minimum(concurrent_chunks)
+                },
+                &array
+                    .codecs()
+                    .recommended_concurrency(&chunk_representation)?,
+            );
+        let codec_options = CodecOptionsBuilder::new()
+            .concurrent_target(codec_concurrent_target)
+            .build();
+
+        // println!("chunks_concurrent_limit {chunks_concurrent_limit:?} codec_concurrent_target {codec_concurrent_target:?}");
+        let n_chunks = usize::try_from(chunks.shape().iter().product::<u64>()).unwrap();
+        // NOTE: Could init memory per split with for_each_init and then reuse it with retrieve_chunk_into_array_view_opt.
+        //       But that might be cheating against tensorstore.
+        rayon_iter_concurrent_limit::iter_concurrent_limit!(
+            chunks_concurrent_limit,
+            (0..n_chunks).into_par_iter(),
+            for_each,
+            |chunk_index| {
+                let chunk_indices = zarrs::array::unravel_index(chunk_index as u64, chunks.shape());
+                // println!("Chunk/shard: {:?}", chunk_indices);
+                let bytes = array
+                    .retrieve_chunk_opt(&chunk_indices, &codec_options)
+                    .unwrap();
+                *bytes_decoded.lock().unwrap() += bytes.len();
+            }
+        );
     }
     let bytes_decoded = bytes_decoded.into_inner()?;
     let duration = SystemTime::now().duration_since(start)?.as_secs_f32();