+++ CHANGELOG.md
@@ -7,8 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+ - Add benchmark data generator: `scripts/generate_benchmark_array.py`
+ - Add benchmark runners: `scripts/run_benchmark/read_{all,chunks}.py`
+ - Add tensorstore benchmark script
+
 ### Changed
  - Bump `cargo-dist` to to 0.9.0 and use GitHub `macos-14` runners for `aarch64-apple-darwin`
+ - `zarrs_benchmark_read_{sync,async}`, `zarrs_binary2zarr`, `zarrs_reencode`
+   - Make `--concurrent-chunks` optional, choosing optimal by default
+ - `zarrs_ncvar2zarr`
+   - Change `--num-parallel-blocks` to `--concurrent-blocks` and make it optional
+   - Remove `--no-parallel-codecs`
+ - Improve internal concurrency
+ - Update benchmarks with `zarrs` v0.12.0
+ - Update dependencies
 
 ## [0.2.3] - 2024-02-06